+++ tests/testthat/test-linters.R
@@ -8,35 +8,17 @@ test_that("linter warns about absolute paths and relative paths", {
 
   absPathLintedIndices <- result[["server.R"]]$absolute.paths$indices
   expect_identical(as.numeric(absPathLintedIndices), 15)
-
-  relPathLintedIndices <- result[["server.R"]]$invalid.relative.paths$indices
-  expect_identical(as.numeric(relPathLintedIndices), 16)
 })
-
-test_that("badRelativePaths identifies bad paths correctly", {
-
-  path <- "R/test.R"
-  ok <- "file.path('../inst/include')"
-  expect_false(badRelativePaths(ok, path = path))
-
-  bad <- "file.path('../../elsewhere')"
-  expect_true(badRelativePaths(bad, path = path))
-
-  ok <- "'../foo', '../bar', '../baz'"
-  expect_false(badRelativePaths(ok, path = path))
-
-})
-
 
 test_that("The linter identifies files not matching in case sensitivity", {
-  result <- lint("shinyapp-with-absolute-paths")
+  result <- lint(test_path("shinyapp-with-absolute-paths"))
   server.R <- result[["server.R"]]
   filepath.capitalization <- server.R[["filepath.capitalization"]]
   expect_equal(as.integer(filepath.capitalization$indices), 31)
 })
 
 test_that("The linter identifies files with Markdown links not matching in case sensitivity", {
-  result <- lint("test-rmd-bad-case")
+  result <- lint(test_path("test-rmd-bad-case"))
   index.Rmd <- result[["index.Rmd"]]
   filepath.capitalization <- index.Rmd[["filepath.capitalization"]]
   expect_equal(as.integer(filepath.capitalization$indices), 29)